Finish the Clang/LLVM 19 lane; park the cross-TU gate for B
E, on the Lint.cmake registry: the pinned image moves to Debian trixie
(LLVM 19) and the gating profile gains readability-redundant-casting; the
85 redundant casts it finds are removed. Clang-19-promoted legacy idioms the
corpus preserves are demoted to compat-noise in cmake/Lint.cmake
(cast-function-type mismatches, Info-ZIP K&R definitions, VC6 implicit
copies; SGP implicit declarations and callback pointers stay report-only as
vendor behavior, and the vendor profile promotes recovery warnings to errors
only in the gating lane). clang-tidy runs only over reccmp source roots, so
the pristine zlib/Info-ZIP statics in the compile database keep upstream
silence. Pin the reccmp revision with the LLVM 19 collector.

Claims in check-identities are scoped by (target, address): extension DLLs
legitimately reuse RVAs.

B: the cross-TU consistency gate stays tested but uncalled (TODO(B)); its
remaining hits are the legal extern T[]/T[N] completion idiom.

+++ src/wiz8/engine_code/GDCamera.cpp
@@ -490,14 +490,14 @@ unsigned char GDCamera::ComputeTrackingOrientation(
         (GetElevationAngle(&m_position_08c, target)
          + g_camera_angle_period_005ec014)
         - (m_pitch + g_camera_angle_period_005ec014);
-    if ((double)fabs((double)angle_delta) > g_camera_pi_005ec2a0) {
+    if (fabs((double)angle_delta) > g_camera_pi_005ec2a0) {
         if (angle_delta >= 0.0f) {
             angle_delta -= g_camera_angle_period_005ec014;
         } else {
             angle_delta += g_camera_angle_period_005ec014;
         }
     }
-    if ((double)fabs((double)pitch_delta) > g_camera_pi_005ec2a0) {
+    if (fabs((double)pitch_delta) > g_camera_pi_005ec2a0) {
         if (pitch_delta >= 0.0f) {
             pitch_delta -= g_camera_angle_period_005ec014;
         } else {
